requirements.txt
fastapi
pydantic
Scrapy
python-dotenv
certifi
google-books-api-wrapper
elasticsearch